Initial commit
* simple service with three endpoints (2 for graphql and upload)
* provides access to the SM engine ES index and DB in the form of GraphQL queries
