+++ package.json
@@ -3,11 +3,14 @@
   "version": "0.0.1",
   "description": "sm-engine GraphQL interface",
   "main": "server.js",
-  "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1",
-    "start": "node server.js"
-  },
   "author": "Artem Tarasov <artem.tarasov@embl.de>",
+  "scripts": {
+    "start": "node server.js",
+    "build": "webpack -p",
+    "build-dev": "webpack -d",
+    "dev": "webpack -d -w",
+    "watch": "webpack -p -w"
+  },
   "license": "ISC",
   "dependencies": {
     "body-parser": "^1.15.2",
@@ -19,12 +22,15 @@
     "graphql-server-express": "^0.4.3",
     "graphql-tools": "^0.8.3",
     "knex": "^0.12.6",
+    "multer": "^1.2.1",
     "node-fetch": "^1.6.3",
     "pg": "^6.1.0",
     "pg-promise": "^5.4.7",
     "sprintf-js": "^1.0.3"
   },
   "devDependencies": {
+    "babel-cli": "^6.18.0",
+    "babel-preset-es2015": "^6.18.0",
     "nodemon": "^1.11.0"
   }
 }
